Move login URL check to the trait.
